# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  dto.AvailableTransitionsView:
    properties:
      current_state:
        $ref: '#/definitions/dto.StateView'
      transitions:
        items:
          $ref: '#/definitions/dto.TransitionView'
        type: array
    type: object
...
  dto.CreateProductRequest:
    properties:
      allow_backorder:
        type: boolean
      attributes:
        items:
          $ref: '#/definitions/dto.ProductAttributeInput'
        type: array
      barcode:
        type: string
      brand_id:
        type: integer
      category_id:
        type: integer
      channels:
        items:
          type: string
        type: array
      colors:
        items:
          type: string
        type: array
      compare_at_price:
        minimum: 0
        type: number
      cost:
        minimum: 0
        type: number
      description:
        type: string
      images:
        items:
          type: string
        type: array
      is_digital:
        type: boolean
      is_new:
        type: boolean
      low_stock_threshold:
        type: integer
      meta_description:
        type: string
      meta_title:
        type: string
      name:
        maxLength: 255
        minLength: 3
        type: string
      price:
        minimum: 0
        type: number
      published_at:
        type: string
      sizes:
        items:
          type: string
        type: array
      sku:
        maxLength: 50
        minLength: 3
        type: string
      status:
        enum:
        - draft
        - active
        - inactive
        - archived
        type: string
      stock:
        minimum: 0
        type: integer
      store_id:
        type: integer
      tags:
        items:
          type: string
        type: array
      track_inventory:
        type: boolean
      visibility:
        type: string
      warehouse_location:
        type: string
      weight:
        minimum: 0
        type: number
    required:
    - name
    - price
    - sku
    type: object
...
  dto.ProductAttributeResponse:
    properties:
      name:
        type: string
      values:
        items:
          type: string
        type: array
    type: object
...
  dto.ProductWithLike:
    properties:
      allow_backorder:
        type: boolean
      attributes:
        items:
          $ref: '#/definitions/dto.ProductAttributeResponse'
        type: array
      barcode:
        type: string
      brand:
        $ref: '#/definitions/dto.BrandResponse'
      brand_id:
        type: integer
      category:
        $ref: '#/definitions/dto.CategoryResponse'
      category_id:
        type: integer
      channels:
        items:
          type: string
        type: array
      colors:
        items:
          type: string
        type: array
      compare_at_price:
        type: number
      cost:
        type: number
      created_at:
        type: string
      description:
        type: string
      id:
        type: integer
      images:
        items:
          type: string
        type: array
      is_digital:
        type: boolean
      is_liked:
        type: boolean
      is_new:
        type: boolean
      low_stock_threshold:
        type: integer
      meta_description:
        type: string
      meta_title:
        type: string
      name:
        type: string
      price:
        type: number
      published_at:
        type: string
      rating:
        type: number
      reviews_count:
        type: integer
      sizes:
        items:
          type: string
        type: array
      sku:
        type: string
      slug:
        type: string
      status:
        type: string
      stock:
        type: integer
      store:
        $ref: '#/definitions/dto.ProductStoreSummary'
      store_id:
        type: integer
      tags:
        items:
          type: string
        type: array
      track_inventory:
        type: boolean
      updated_at:
        type: string
      visibility:
        description: Publishing extras
        type: string
      warehouse_location:
        type: string
      weight:
        type: number
      workflow_state:
        $ref: '#/definitions/dto.StateView'
    type: object
...
  utils.Response:
    properties:
      code:
        type: integer
      error:
        type: string
      message:
        type: string
      success:
        type: boolean
    type: object
host: localhost:8080
info:
  contact:
    email: support@luxe.com
    name: API Support
  description: Production-grade e-commerce backend
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: Shopping Platform API
  version: "1.0"
paths:
...
  /products:
    get:
      consumes:
      - application/json
      description: Get products with pagination and optional filters
      parameters:
      - default: 20
        description: Items per page (default 20, max 100)
        in: query
        name: limit
        type: integer
      - default: 0
        description: Number of items to skip
        in: query
        name: offset
        type: integer
      - description: Product status
        enum:
        - active
        - draft
        - archived
        in: query
        name: status
        type: string
      - description: Filter by product name
        in: query
        name: name
        type: string
      - description: Filter by SKU
        in: query
        name: sku
        type: string
      - description: Filter by category ID
        in: query
        name: category_id
        type: integer
      - description: Filter by brand ID
        in: query
        name: brand_id
        type: integer
      - description: Minimum price
        in: query
        name: min_price
        type: number
      - description: Maximum price
        in: query
        name: max_price
        type: number
      - description: Minimum rating (0–5)
        in: query
        name: min_rating
        type: number
      - description: Maximum rating
        in: query
        name: max_rating
        type: number
      - description: Minimum review count
        in: query
        name: min_reviews
        type: integer
      - description: Maximum review count
        in: query
        name: max_reviews
        type: integer
      - description: Digital products only
        in: query
        name: is_digital
        type: boolean
      - description: New products only
        in: query
        name: is_new
        type: boolean
      - description: Sort order
        enum:
        - rating_desc
        - rating_asc
        - newest
        - reviews_desc
        - price_asc
        - price_desc
        in: query
        name: sort
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/utils.Response'
            - properties:
                data:
                  properties:
                    limit:
                      type: integer
                    offset:
                      type: integer
                    products:
                      items:
                        $ref: '#/definitions/dto.ProductWithLike'
                      type: array
                    total:
                      type: integer
                  type: object
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.Response'
      summary: List products
      tags:
      - Products
    post:
      consumes:
      - application/json
      description: Add a new product to the catalog
      parameters:
      - description: Product details
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.CreateProductRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            allOf:
            - $ref: '#/definitions/utils.Response'
            - properties:
                data:
                  $ref: '#/definitions/dto.ProductResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.Response'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/utils.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.Response'
      security:
      - BearerAuth: []
      summary: Create product
      tags:
      - Products
  /products/{id}:
    delete:
      consumes:
      - application/json
      description: Soft delete a product by its ID
      parameters:
      - description: Product ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/utils.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.Response'
      security:
      - BearerAuth: []
      summary: Delete product
      tags:
      - Products
    get:
      consumes:
      - application/json
      description: Fetch a single product using either its numeric ID or slug
      parameters:
      - description: Product identifier (ID or slug)
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/utils.Response'
            - properties:
                data:
                  properties:
                    is_liked:
                      type: boolean
                    product:
                      $ref: '#/definitions/dto.ProductResponse'
                    stock_subscribed:
                      type: boolean
                  type: object
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/utils.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.Response'
      summary: Get product by ID or slug
      tags:
      - Products
    put:
      consumes:
      - application/json
      description: Update an existing product by its ID
      parameters:
      - description: Product ID
        in: path
        name: id
        required: true
        type: integer
      - description: Product update data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.UpdateProductRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/utils.Response'
            - properties:
                data:
                  $ref: '#/definitions/dto.ProductResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/utils.Response'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/utils.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.Response'
      security:
      - BearerAuth: []
      summary: Update product
      tags:
      - Products
...
  /products/{id}/available-transitions:
    get:
      description: Returns available workflow events from the product's current state
        (approve, publish, discontinue, etc.).
      parameters:
      - description: Product ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/utils.Response'
            - properties:
                data:
                  $ref: '#/definitions/dto.AvailableTransitionsView'
              type: object
      security:
      - BearerAuth: []
      summary: List product workflow transitions (admin)
      tags:
      - Products
...
  /products/{id}/related:
    get:
      consumes:
      - application/json
      description: Fetch products from the same category, ordered by rating and review
        count
      parameters:
      - description: Product ID
        in: path
        name: id
        required: true
        type: integer
      - description: Number of related products (default 4, max 10)
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/utils.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/dto.ProductResponse'
                  type: array
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/utils.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.Response'
      summary: Get related products
      tags:
      - Products
...
  /products/{id}/transition:
    post:
      consumes:
      - application/json
      description: Fires a workflow event such as approve, publish, discontinue, or
        archive. Updates product status via the engine mirror.
      parameters:
      - description: Product ID
        in: path
        name: id
        required: true
        type: integer
      - description: Workflow event
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.PerformProductTransitionRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/utils.Response'
            - properties:
                data:
                  $ref: '#/definitions/dto.ProductTransitionResponse'
              type: object
      security:
      - BearerAuth: []
      summary: Transition product workflow state (admin)
      tags:
      - Products
  /products/bulk:
    delete:
      consumes:
      - application/json
      description: Soft delete multiple products by their IDs (admin only)
      parameters:
      - description: Product IDs to delete
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.BulkDeleteProductsRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.Response'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/utils.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/utils.Response'
      security:
      - BearerAuth: []
      summary: Bulk delete products
      tags:
      - Products
    post:
      consumes:
      - application/json
      description: Create multiple products in a single request (admin only)
      parameters:
      - description: Array of products
        in: body
        name: request
        required: true
        schema:
          items:
            $ref: '#/definitions/dto.CreateProductRequest'
          type: array
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            allOf:
            - $ref: '#/definitions/utils.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/dto.ProductResponse'
                  type: array
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.Response'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/utils.Response'
      security:
      - BearerAuth: []
      summary: Bulk create products
      tags:
      - Products
  /products/suggestions:
    post:
      consumes:
      - application/json
      description: Returns products from the same categories as the provided product
        IDs, excluding the products themselves. Useful for cart page "you might also
        like".
      parameters:
      - description: Product IDs and limit
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.SuggestionsRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Suggestions fetched
          schema:
            allOf:
            - $ref: '#/definitions/utils.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/dto.ProductResponse'
                  type: array
              type: object
        "400":
          description: Invalid request
          schema:
            $ref: '#/definitions/utils.Response'
        "500":
          description: Internal error
          schema:
            $ref: '#/definitions/utils.Response'
      security:
      - BearerAuth: []
      summary: Get smart product suggestions
      tags:
      - Products
...
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and the JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"